+++ CHANGELOG.md
@@ -26,6 +26,11 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian/report/compare/release-3.5.0...master
 
+### Added
+- Add `failedActions` property to `Verdict`. Resolve [JPERF-537].
+
+[JPREF-537]: https://ecosystem.atlassian.net/browse/JPERF-537
+
 ## [3.5.0] - 2019-07-05
 [3.5.0]: https://github.com/atlassian/report/compare/release-3.4.0...release-3.5.0